Move `workspace_storage` into `storage` sub-module
